pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "deeponet-pricing"
version = "0.1.0"
description = "DeepONet-based option pricing for Heston IV and rBergomi models"
requires-python = ">=3.11"
dependencies = [
    "deepxde>=1.15.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "torch>=2.0",
    "matplotlib>=3.7",
    "mlflow>=2.10",
    "py_lets_be_rational>=1.0.1",
    "pyyaml>=6.0",
    "tqdm>=4.60",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[project.scripts]
deeponet-pricing = "deeponet_pricing.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: long-running tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
